feat: removed mechanism to update arbitration fees
Now if the arbitration fee increases between the time the arbitration request is made on the Foreign Chain and the time Realitio is notified of that request on the Home Chain, the dispute will not be created.

This mitigates a potential vector for attack when the arbitration fees are about to go up: a bad actor, who wants A to win, stakes on B and then requests arbitration in favour of B, so the proxy would wait to notify Realitio. Then once the fees have increased, s/he stakes on A, but does not pay for the remaining arbitration fees. If no one else pays for the fee, A would win and the attacker would get his/her deposit back.